backend/pyproject.toml
[project]
name = "margo-backend"
version = "0.1.0"
description = "Margó — Food cost management API for Belgian restaurants"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",
    "sqlalchemy[asyncio]>=2.0.36,<3.0",
    "asyncpg>=0.30.0,<1.0",
    "alembic>=1.14.0,<2.0",
    "pydantic-settings>=2.6.0,<3.0",
    "PyJWT[crypto]>=2.9.0,<3.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0,<1.0",
    "email-validator>=2.2.0,<3.0",
    "anthropic>=0.40.0,<1.0",
    "lxml>=5.0.0,<7.0",
    "pdfplumber>=0.11.0,<1.0",
    "stripe>=8.0.0,<12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-asyncio>=0.24.0,<1.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
